- Added map to detail view. - Added UI test. - Minor fixes on code

+++ app/src/main/java/com/sergiom/madtivities/data/entities/MadEventItem.kt
@@ -31,8 +31,8 @@ data class MadEventItem(
     )
 
     data class EventLocation (
-        val latitude: Float,
-        val longitude: Float
+        val latitude: Double,
+        val longitude: Double
     )
 
     data class EventOrganization(
